reset image cache button, success banners, reload notification posting

+++ Gametime/GTSettingsViewController.m
@@ -9,12 +9,25 @@
 #import "GTSettingsViewController.h"
 #import "GTTeamObject.h"
 #import "GTPlayerObject.h"
+#import "CWStatusBarNotification.h"
+#import "SDWebImageManager.h"
+#import "GTMainViewController.h"
 
 @interface GTSettingsViewController ()
 
 @end
 
 @implementation GTSettingsViewController
+
+- (instancetype)init {
+    self = [super initWithStyle:UITableViewStyleGrouped];
+    
+    if (self) {
+        
+    }
+    
+    return self;
+}
 
 - (void)viewDidLoad {
     [super viewDidLoad];
@@ -29,6 +42,7 @@
 #pragma mark - actions
 
 - (void)doneBarButtonItemTapped {
+    [[NSNotificationCenter defaultCenter] postNotificationName:kGametimeRefreshTeamNotificationName object:nil];
     [self dismissViewControllerAnimated:YES completion:NULL];
 }
 
@@ -39,7 +53,7 @@
 }
 
 - (NSInteger)tableView:(UITableView *)tableView numberOfRowsInSection:(NSInteger)section {
-    return 1;
+    return 2;
 }
 
 - (CGFloat)tableView:(UITableView *)tableView heightForRowAtIndexPath:(nonnull NSIndexPath *)indexPath {
@@ -52,9 +66,10 @@
     if (!cell) {
         cell = [[UITableViewCell alloc] initWithStyle:UITableViewCellStyleDefault reuseIdentifier:kGametimeSettingsResetCellIdentifier];
         cell.textLabel.textColor = [UIColor colorWithRed:0.864 green:0.148 blue:0.157 alpha:1.00];
-        cell.textLabel.text = @"Reset Gametime Data";
         cell.accessoryType = UITableViewCellAccessoryDisclosureIndicator;
     }
+    
+    cell.textLabel.text = indexPath.row == 0 ? @"Reset Gametime Data" : @"Reset Image Cache";
     
     return cell;
 }
@@ -62,8 +77,23 @@
 - (void)tableView:(UITableView *)tableView didSelectRowAtIndexPath:(NSIndexPath *)indexPath {
     [tableView deselectRowAtIndexPath:indexPath animated:YES];
     
-    [GTTeamObject deleteAll];
-    [GTPlayerObject deleteAll];
+    CWStatusBarNotification *successNotification = [[CWStatusBarNotification alloc] init];
+    successNotification.notificationLabelBackgroundColor = [UIColor colorWithRed:0.182f green:0.613f blue:0.312f alpha:1.00f];
+    successNotification.notificationLabelTextColor = [UIColor whiteColor];
+
+    if (indexPath.row == 0) {
+        [GTTeamObject deleteAll];
+        [GTPlayerObject deleteAll];
+        [successNotification displayNotificationWithMessage:@"✓ Reset Gametime Data" forDuration:1.5];
+    }
+    
+    else {
+        [[SDWebImageManager sharedManager] cancelAll];
+        [[SDImageCache sharedImageCache] clearMemory];
+        [[SDImageCache sharedImageCache] clearDiskOnCompletion:^{
+            [successNotification displayNotificationWithMessage:@"✓ Reset Image Cache" forDuration:1.5];
+        }];
+    }
 }
 
 @end
